+++ tools/build-manual.R
@@ -2,6 +2,7 @@
 local({
   stopifnot(file.exists("DESCRIPTION"), dir.exists("man"))
   source("tools/check-documentation.R", local = TRUE)
+  source("tools/check-dependencies.R", local = TRUE)
   dir.create("output/pdf", recursive = TRUE, showWarnings = FALSE)
   status <- system2(file.path(R.home("bin"), "R"), c(
     "CMD", "Rd2pdf", "--no-preview", "--force",